Update AliensTemplateCode.html
New changes to the file include another tab expected in the navigation bar of the website that is required for this part of the project. Finally decided to use external CSS stylesheet to make the code much more efficient, look more organized and the styles to be more universal. Navigation bar is still subject to change, might use the unordered list method as a substitute and will also start using different fonts soon but this is the progress so far. Also changed some content.