Update style.css
Finally imported new fonts from the Google Fonts to the stylesheet, namely, Raleway, Roboto, and Inter Tight. To change the look and style of the text display to make it look smoother and more stylish than the default text. Aside from that, I've also made the header bar and navigation menu sticky to the top.