Create HeatblastPage(MajorWorkInProgress).html
There are still a few things to add here, like information, history and picture of the character. But so far, this is info present for Heatblast As you can see, I have decided to use different fonts now for the headers and the content to stylize the webpage more and decided to make the header and navigation bar sticky to the top. But through a few experiments and using multiple copy-pastes of Lorem Ipsum, I have discovered that after a certain length of scrolling, the div of the header and nav bar unsticks and disappears from the page. Still going to find a solution to this as this is not the intended feature. But until then, this is the current project's look.