Update and rename HeatblastPage.html to Heatblast.html
Added a lot of stuff today. Stuff I added was a further description about Heatblast, and a new section elaborating on his abilities inside a table, accompanied by an image at the left of him using his fire powers. Fixed an issue concerning the bottom half of the navigation bar being unclickable, solved it by changing the name of the div parenting it, because apparently, I forgot to change it. Also added an icon for the page already, which is a small image of the omnitrix. Aside from that, that's most of this page done. I just need to add the appearances throughout each edition of the series and his weaknesses, and the site should be as good as done.